# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Identity Link DB Client API'
  description: 'API documentation for Identity Link DB Clients'
  version: '1.0'
servers:
  -
    url: /clients
    description: 'Clients API base path'
paths:
...
  '/api/v1/secret/{id}':
    get:
      tags:
        - Secret
      summary: 'Fetch a secret by ID'
      operationId: e58aa4c5257e6adaf1725fa00a868607
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Secret data'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { secret: { $ref: '#/components/schemas/Secret' } }, type: object }
                type: object
        '404':
          description: 'Secret not found'
    put:
      tags:
        - Secret
      summary: 'Update a secret'
      operationId: 66c6bc194517834467470e74919de24c
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Secret'
      responses:
        '200':
          description: 'Secret updated'
        '400':
          description: 'Validation error'
        '404':
          description: 'Secret not found'
    delete:
      tags:
        - Secret
      summary: 'Delete a secret'
      operationId: 34891b38848eaa843ba394e3bd030c94
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 'Secret deleted'
        '404':
          description: 'Secret not found'
  /api/v1/secret:
    post:
      tags:
        - Secret
      summary: 'Create a new secret'
      operationId: 8879a9a5a5308c0868311de389e06560
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Secret'
      responses:
        '201':
          description: 'Secret created (password omitted)'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { secret: { $ref: '#/components/schemas/Secret' } }, type: object }
                type: object
        '400':
          description: 'Validation error'
  /api/v1/secret/issue:
    post:
      tags:
        - Secret
      summary: 'Issue a new secret (auto-generates password)'
      operationId: 20fbd80731f67775bdfda89455d9565b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                expirationPeriod:
                  description: 'Relative expiration period'
                  type: string
                  enum: [1d, 1w, 1m, 3m, 9m, 1y, 2y]
                client:
                  $ref: '#/components/schemas/Client'
                passwordHint:
                  description: 'Optional hint to help recall the password'
                  type: string
                  maxLength: 500
              type: object
      responses:
        '201':
          description: 'Secret issued (password omitted)'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { secret: { $ref: '#/components/schemas/Secret' } }, type: object }
                type: object
        '400':
          description: 'Validation error'
components:
  schemas:
...
    Client:
      title: Client
      description: 'OAuth2 Client entity used for identity link authorization'
      properties:
        id:
          description: 'Client UUID'
          type: string
          format: uuid
        name:
          description: 'Unique client name'
          type: string
          maxLength: 100
        description:
          description: 'Client description'
          type: string
          maxLength: 3000
        redirectUri:
          description: 'Client redirect URIs'
          type: array
          items:
            type: string
            format: uri
            maxLength: 3000
        grantTypes:
          description: 'Allowed OAuth2 grant types'
          type: array
          items:
            type: string
            enum:
              - client_credentials
              - password
              - authorization_code
              - refresh_token
              - implicit
        scopes:
          description: 'OAuth2 scopes for the client'
          type: array
          items:
            type: string
        groups:
          description: 'Group IDs assigned to the client'
          type: array
          items:
            type: string
        secrets:
          description: 'Secret keys associated with the client'
          type: array
          items:
            type: string
        isPublic:
          description: 'Indicates whether the client is public'
          type: boolean
        isSystem:
          description: 'Indicates whether the client is managed by the system and cannot be updated or deleted through the API'
          type: boolean
          readOnly: true
      type: object
...
    Secret:
      title: Secret
      description: "Represents a client's secret, including password hint and expiration info."
      properties:
        id:
          description: 'Unique identifier of the secret'
          type: string
          format: uuid
        password:
          description: 'Plaintext password (write-only)'
          type: string
          maxLength: 50
          writeOnly: true
        passwordHint:
          description: 'Hint to help recall the password'
          type: string
          maxLength: 500
        expirationDateTime:
          description: 'Expiration timestamp for the secret'
          type: string
          format: date-time
        client:
          $ref: '#/components/schemas/Client'
        isSystem:
          description: 'Indicates whether the secret is managed by the system and cannot be updated or deleted through the API'
          type: boolean
          readOnly: true
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      description: 'JWT Bearer token authorization'
      bearerFormat: JWT
      scheme: bearer
security:
  -
    bearerAuth: []
...